Big cleanup of architecture. Now its well organized and easier to access
