+++ src/App.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
-import Homepage from './components/Homepage';
-import LoginPage from './components/LoginPage';
-import ActusMainPage from './components/ActusMainPage';
-import OfferPage from './components/OfferPage';
-import Offer from './components/Offer';
-import OfferRate from './components/OfferRate';
-import ActusPage from './components/ActusPage';
-import SubscriberPage from './components/SubscriberPage';
-import SubscriberWelcomePage from './components/SubscriberWelcomePage';
+import Homepage from './pages/Homepage';
+import LoginPage from './pages/LoginPage';
+import ActusPage from './pages/ActusPage';
+import OfferPage from './pages/OfferPage';
+import Offer from './offer/Offer';
+import OfferRate from './offer/OfferRate';
+import RecipientLocation from './offer/RecipientLocation';
+import Actus from './components/Actus';
+import SubscriberPage from './pages/SubscriberPage';
+import SubscriberWelcomePage from './subscriber/SubscriberWelcomePage';
+import SubscriberHasRecipients from './subscriber/SubscriberHasRecipients';
+import Contact from './components/ContactPage';
 import FormPage from './pages/FormPage';
 import FormSignUpStep1 from './formSignUp/FormSignUpStep1';
 import FormSignUpStep2 from './formSignUp/FormSignUpStep2';
-import SubscriberHasRecipients from './components/SubscriberHasRecipients';
-import RecipientLocation from './components/RecipientLocation';
-import Contact from './components/ContactPage';
 
 import './App.css';
 
@@ -24,7 +24,7 @@ function App() {
     <>
       <Routes>
         <Route path="/" element={<Homepage />} />
-        <Route path="/actus" element={<ActusMainPage />} />
+        <Route path="/actus" element={<ActusPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/createaccount" element={<FormPage />}>
           <Route path="step1" element={<FormSignUpStep1 />} />
@@ -38,7 +38,7 @@ function App() {
         <Route path="/subscribers" element={<SubscriberPage />}>
           <Route path="welcome" element={<SubscriberWelcomePage />} />
           <Route path="recipients" element={<SubscriberHasRecipients />} />
-          <Route path="actionsSolidaires" element={<ActusPage />} />
+          <Route path="actionsSolidaires" element={<Actus />} />
           <Route path="contact" element={<Contact />} />
         </Route>
       </Routes>
