--- src/components/style/SubscriberContainer.module.scss
@@ -1,5 +1,0 @@
-.subscriberDownContainer {
-  margin-top: 8vh;
-  display: flex;
-  flex-direction: row;
-}
