--- src/components/style/SubscriberRightMainContainer.module.scss
@@ -1,6 +1,0 @@
-.subscriberRightMainContainer {
-  width: 70%;
-  height: 300vh;
-  display: flex;
-  flex-direction: column;
-}
