+++ src/formSignUp/style/FormSignUpStep1.module.scss
