+++ src/formSignUp/style/FormSignUpStep2.module.scss
