+++ src/homepage/HomepageConcept.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ConceptCard from './ConceptCard';
 import ConceptButtons from './ConceptButtons';
 
+import style from './style/HomepageConcept.module.scss';
 import MyImgConcept from '../assets/MyImgConcept.png';
 import MyImgConcept2 from '../assets/MyImgConcept2.png';
 import MyImgConcept3 from '../assets/MyImgConcept3.png';
@@ -11,8 +12,6 @@ import Optique from '../assets/optique.png';
 import FraisMedicaux from '../assets/fraisMedicaux.png';
 import Dentaire from '../assets/dentaire.png';
 import Teleconsultation from '../assets/teleconsultation.png';
-
-import style from './style/HomepageConcept.module.scss';
 
 function HomepageConcept() {
   const Concepts = [
