+++ src/homepage/style/ConceptCard.module.scss
