+++ src/offer/style/Offer.module.scss
