+++ src/offer/style/OfferRate.module.scss
