+++ src/offer/style/RecipientLocation.module.scss
