+++ src/pages/ActusPage.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 
-import Header from './Header';
-import ActusPage from './ActusPage';
+import Header from '../components/Header';
+import Actus from '../components/Actus';
 
 import style from './style/ActusPage.module.scss';
 
-function ActusMainPage() {
+function ActusPage() {
   return (
     <div className={style.actusPageWrapper}>
       <Header />
-      <ActusPage />
+      <Actus />
     </div>
   );
 }
 
-export default ActusMainPage;
+export default ActusPage;
