+++ src/pages/OfferPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 
-import Header from './Header';
+import Header from '../components/Header';
 
 function OfferPage() {
   return (
